backend/requirements.txt
# API Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Auth & Security
PyJWT[crypto]==2.9.0
google-auth==2.35.0

# Firebase
firebase-admin==6.5.0

# HTTP Client (for calling ingest-service)
httpx==0.27.0

# GitHub
PyGithub==2.1.1

# Utils
python-dotenv==1.0.0
pydantic==2.9.0